[PACKAGES-MIGRATION] move version
