+++ packages/base/src/commands/git-metadata/upload.ts
@@ -12,7 +12,7 @@ import {Logger, LogLevel} from '../../helpers/logger'
 import {MetricsLogger, getMetricsLogger} from '../../helpers/metrics'
 import {UploadStatus} from '../../helpers/upload'
 import {getRequestBuilder, timedExecAsync} from '../../helpers/utils'
-import {cliVersion} from '../../helpers/version'
+import {cliVersion} from '../../version'
 
 import {apiHost, datadogSite, getBaseIntakeUrl} from './api'
 import {getCommitInfo, newSimpleGit} from './git'
