--- packages/base/src/helpers/version.ts
@@ -1,2 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-var-requires
-export const cliVersion = require('../../../datadog-ci/package.json').version as string
