+++ packages/base/src/version.ts
