+++ packages/datadog-ci/src/commands/dsyms/upload.ts
@@ -13,11 +13,10 @@ import {upload, UploadStatus} from '@datadog/datadog-ci-base/helpers/upload'
 import {buildPath, getRequestBuilder, resolveConfigFromFileAndEnvironment} from '@datadog/datadog-ci-base/helpers/utils'
 import * as validation from '@datadog/datadog-ci-base/helpers/validation'
 import {checkAPIKeyOverride} from '@datadog/datadog-ci-base/helpers/validation'
+import {cliVersion} from '@datadog/datadog-ci-base/version'
 import chalk from 'chalk'
 import {Command, Option} from 'clipanion'
 import upath from 'upath'
-
-import {cliVersion} from '../../version'
 
 import {CompressedDsym, Dsym, DWARF} from './interfaces'
 import {
