+++ packages/datadog-ci/src/commands/elf-symbols/__tests__/upload.test.ts
@@ -9,9 +9,8 @@ import {
   MultipartStringValue,
   UploadStatus,
 } from '@datadog/datadog-ci-base/helpers/upload'
+import {cliVersion} from '@datadog/datadog-ci-base/version'
 import upath from 'upath'
-
-import {cliVersion} from '../../../version'
 
 import {ElfClass} from '../elf-constants'
 import {uploadMultipartHelper} from '../helpers'
