+++ packages/datadog-ci/src/commands/lambda/flare.ts
@@ -35,11 +35,10 @@ import {requestConfirmation, requestFilePath} from '@datadog/datadog-ci-base/hel
 import * as helpersRenderer from '@datadog/datadog-ci-base/helpers/renderer'
 import {renderAdditionalFiles, renderProjectFiles} from '@datadog/datadog-ci-base/helpers/renderer'
 import {formatBytes} from '@datadog/datadog-ci-base/helpers/utils'
+import {cliVersion} from '@datadog/datadog-ci-base/version'
 import chalk from 'chalk'
 import {Command, Option} from 'clipanion'
 import upath from 'upath'
-
-import {cliVersion} from '../../version'
 
 import {
   AWS_DEFAULT_REGION_ENV_VAR,
