+++ packages/datadog-ci/src/commands/sourcemaps/upload.ts
@@ -18,11 +18,10 @@ import {getMetricsLogger, MetricsLogger} from '@datadog/datadog-ci-base/helpers/
 import {upload, UploadStatus} from '@datadog/datadog-ci-base/helpers/upload'
 import {getRequestBuilder, buildPath} from '@datadog/datadog-ci-base/helpers/utils'
 import * as validation from '@datadog/datadog-ci-base/helpers/validation'
+import {cliVersion} from '@datadog/datadog-ci-base/version'
 import chalk from 'chalk'
 import {Command, Option} from 'clipanion'
 import upath from 'upath'
-
-import {cliVersion} from '../../version'
 
 import {Sourcemap} from './interfaces'
 import {
