+++ packages/datadog-ci/src/commands/unity-symbols/upload.ts
@@ -21,10 +21,9 @@ import {
 } from '@datadog/datadog-ci-base/helpers/utils'
 import * as validation from '@datadog/datadog-ci-base/helpers/validation'
 import {checkAPIKeyOverride} from '@datadog/datadog-ci-base/helpers/validation'
+import {cliVersion} from '@datadog/datadog-ci-base/version'
 import {Command, Option} from 'clipanion'
 import upath from 'upath'
-
-import {cliVersion} from '../../version'
 
 import * as dsyms from '../dsyms/upload'
 import {createUniqueTmpDirectory} from '../dsyms/utils'
